chore(Restructure project directories): rearranged project folders
