+++ .eslintignore
@@ -1,0 +1,7 @@
+src/test
+node_modules
+coverage
+bin
+.nyc_output
+build
+
