+++ package.json
@@ -3,10 +3,11 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "prebuild": "eslint src/main/**/*.js",
-    "lint:fix": "eslint src/main/**/*.js --fix",
+    "lint": "eslint src/**/*.js",
+    "prebuild": "npm run lint",
+    "lint:fix": "npm run lint -- --fix",
     "clean": "rimraf build/",
-    "build-babel": "babel src/main -d build/src/main",
+    "build-babel": "babel src -d build/src",
     "build": "npm run clean && npm run build-babel",
     "start": "npm run build && node ./bin/server",
     "test": "cross-env NODE_ENV=test nyc mocha \"src/test/**/*.js\"",
