+++ src/models/Entry.js
